+++ modules/cloud_backup_schedule/versions.tf
@@ -1,0 +1,14 @@
+terraform {
+  required_version = ">= 1.10"
+  required_providers {
+    mongodbatlas = {
+      source  = "mongodb/mongodbatlas"
+      version = "~> 2.1"
+    }
+  }
+
+  provider_meta "mongodbatlas" {
+    module_name    = "cluster/cloud_backup_schedule"
+    module_version = "local"
+  }
+}